===== Lauffähige Version===== Der ToolbarButton leitet mit Hilfe der ToolbarBaseActivity von allen Views auf die ProfileActivity weiter Die Thesis erzeugt jetzt zuverlässig eine Datenbank und diese kann auch aus student Sicht verändert werden Die Nutzer supervisor und student werden jetzt immer in der Datenbank erzeugt und hinterlegt SupervisorBoardActivity zeigt jetzt Betreuer aus der Datenbank an. Ein click hierauf leitet den User zu SupervisorViewFromStudentActivity weiter Supervisoren können jetzt thesis zugewiesen werden.

+++ app/src/main/java/com/dlbcsemse/iuthesisconnect/model/UserProfile.kt
@@ -2,8 +2,8 @@ package com.dlbcsemse.iuthesisconnect.model
 import com.dlbcsemse.iuthesisconnect.model.DashboardUserType
 import org.intellij.lang.annotations.Language
 import java.io.Serializable
-data class UserProfile (val id : Long, val name : String, val eMail : String, val type : Int)  {
-    var userId : Long = id
+data class UserProfile (val id : Int, val name : String, val eMail : String, val type : Int)  {
+    var userId : Int = id
     var userName : String = name
     var userEmail : String = eMail
     var userType : DashboardUserType = DashboardUserType.entries[type]
@@ -12,7 +12,7 @@ data class UserProfile (val id : Long, val name : String, val eMail : String, va
     var languages : ArrayList<Language> = ArrayList<Language>()
     var status : AvailabilityStatus = AvailabilityStatus.free
 
-    constructor(id : Long, name : String, eMail : String, type : String)
+    constructor(id : Int, name : String, eMail : String, type : String)
             : this(
         id,
         name,
